+++ HISTORY.md
@@ -2,8 +2,10 @@
 
 -   duplicate use as: ```fofa duplicate -o data.csv -d ip,host,domain -o duplicate.csv```
 -   active of target use as: ```fofa active -target baidu.com,fofa.info``` 
--   active of file use as: ```fofa active -i target.txt``` 
+-   active of file use as: ```fofa active -i target.txt```
 -   active of pipline use as: ```fofa search -f link -s 3 port=80 | fofa active``` 
+-   search of active use as: ```fofa search --isActive port=80```
+
 ## v0.2.19 add inFile in search command
 
 -   use as: ```fofa -f host -uniqByIP -outFile b.cvs -rate 5 -inFile a.cvs```